requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
Pillow==7.1.1
psycopg2==2.8.5
PyJWT==1.7.1
python-slugify==4.0.0
pytz==2019.3
requests==2.23.0
slugify==0.0.1
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.9
